requirements.txt
python-telegram-bot==20.0a6
Pillow==9.5.0
